'prediction.feature' was failing for me: fix it.
Entering "tomorrow" as a time seems to set the time at 12:00 tomorrow,
which won't necessarily return "in 1 day". Setting "in 24 hours" will.

(I'm not sure this is a good commit, maybe the test was right and the
code should be changed? Someone who knows the codebase should look at
it.)