Change to ActiveRecordStore sessions.
Cookie session data is effectively passed in plaintext to the user. It's
signed, but can be replayed; and I will probably need to make sure that
users can't get asked the same question twice anyway, but it just seems
safer not to have session data in cookies.

I don't think this change will break anything, but it's hard to be sure.
I did `ack -w session`, but the change that I had to make (for
unspecified reasons) to application_controller, I only found because of
a comment, so it's possible there are other changes I have to make that
aren't commented. Someone who knows rails and the codebase should look
at this.