Add model for credence question generators.
Also included: a bunch of questions. I haven't really looked through,
but I don't actually anticipate using all of the repositories. Only two
are used in the original game (see Databases.txt).

The import task is dumb, and if you run it twice you'll get duplicates.
I suspect it shouldn't be a rake task, but I'm not sure what would be
idiomatic. Maybe a standalone in scripts/?

Questions are not imported yet, there's no table for them.