question.answer0_correct -> question.correct_index
It's easier to reason about "which answer is correct?" than "is answer 0
correct?". In particular, this fixes a bug where answers were only being
considered correct if they were answer 0; I'm not quite sure how I
failed to notice that, but I created it because the logic was more
complicated than it should have been.

With this migration, any questions currently in the database now have the
wrong answer marked correct. Since I'm presumably currently the only
person with any questions in the database, and they aren't ever getting
reused, this seems acceptable.