Fix for a very nasty bug; the previous date validation was very limited, and allowed two huge loopholes:
First of all, it only disallowed dates beyond the year 9999, whereas exceedingly *early* dates posed the exact same problem. To fix this, I added in a check to make sure the deadline wasn't too early either.

Secondly, it only disallowed invalid dates on creation: If you made a prediction with a valid date, you were allowed to afterwards edit the date to whatever you wanted to, without triggering the validation.

Either of these would have been enough for a date which PredictionBook is unable to store to be entered as the deadline for a prediction, with disastrous results.