Change 'answered_correctly' to 'given_answer' in db.
This seems cleaner. Partly because it fits better with the existing
code, but we can always have a method for the one that isn't directly in
the database. But it makes potential SQL queries nicer:
`answered_correctly` is just `given_answer = correct_index`, but
`given_answer` is slightly more complicated in terms of
`answered_correctly` and `correct_index`.