Fix tests for merge.
I need the 'raise' statements after the 'pending', because apparently
pending no longer causes the block to quit?

Changed 'x.should' to 'expect(x).to' syntax in credence game tests,
because it was getting deprecation warnings. The same warnings are still
coming from other tests, but expect() is also used in some places, and I
like expect better.

Lots of other warnings too, which I'm not sure if they indicate a
problem with my setup.