Fix two questions.
I don't particularly like the new version of the commuting workers one,
it's awkward. But at least it's... maybe not totally unambiguous,
because it might mean "which state has a greater percentage of the total
pool of people commuting to work by public transportation", but I think
that's an unlikely reading.