Turn SortedMultiset questions into Sorted.
In the original game, there was no real difference between them. (If
two answers had the same text, the later one would be ignored, but none
of them do have duplicate text.) We've previously been ignoring them.

Changed them to have weight 0.2 because they're all basically the same
question, and IMO not super interesting.

Disabled the one about the Stanley Cup victories because I have trouble
understanding the stuff on wikipedia, but the original data was
definitely wrong, for the Winnipeg Victorias and I'm pretty sure for
others.